# make some file writes asynchronous, then optimize them