@marco-c 2015-03-28
ensureConnected was used a lot in a MIDlet I'm testing, it made one particular interaction way faster.